+++ Dockerfile
@@ -1,0 +1,40 @@
+# Use an official Python runtime as a parent image
+FROM python:3.9-slim
+
+# Create a non-root user
+RUN groupadd -r appuser && useradd -r -g appuser appuser
+
+# Set the working directory in the container
+WORKDIR /app
+
+# Upgrade system packages for security
+RUN apt-get update && apt-get upgrade -y && \
+    apt-get clean && rm -rf /var/lib/apt/lists/*
+
+# Upgrade pip
+RUN pip install --upgrade pip
+
+# Copy requirements file
+COPY requirements.txt .
+
+# Install Python dependencies
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy the source code into the container
+COPY . .
+
+# Change ownership to non-root user
+RUN chown -R appuser:appuser /app
+
+# Switch to non-root user
+USER appuser
+
+# Expose port (will be overridden by Cloud Run)
+EXPOSE 8080
+
+# Health check
+HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
+    CMD python -c "import requests; requests.get('http://localhost:${PORT:-8000}/')" || exit 1
+
+# Run the FastAPI app using the PORT environment variable
+CMD uvicorn main:app --host 0.0.0.0 --port ${PORT:-8000}
