--- backend/.gitignore
@@ -1,124 +1,0 @@
-# General
-.DS_Store
-Thumbs.db
-*.log
-*.tmp
-*.cache
-*.bak
-*.swp
-*.swo
-
-# Node.js
-node_modules/
-npm-debug.log
-yarn-error.log
-package-lock.json
-
-# Python
-__pycache__/
-*.pyc
-*.pyo
-*.pyd
-venv/
-.env
-
-# Java
-*.class
-*.jar
-*.war
-*.ear
-*.iml
-*.ipr
-*.iws
-out/
-*.log
-
-# C/C++
-*.o
-*.obj
-*.exe
-*.dll
-*.so
-*.dylib
-*.a
-*.lib
-*.lo
-
-# Go
-bin/
-*.exe
-*.out
-*.test
-
-# Rust
-target/
-Cargo.lock
-
-# PHP
-vendor/
-*.log
-
-# Ruby
-*.gem
-log/
-tmp/
-.bundle/
-Gemfile.lock
-
-# Swift/Objective-C
-*.xcworkspace
-*.xcodeproj
-*.xcuserstate
-
-# Dotnet
-bin/
-obj/
-*.dll
-*.pdb
-*.exe
-*.app
-
-# Frontend/Web
-bower_components/
-dist/
-public/
-*.min.js
-*.map
-
-# macOS
-*.DS_Store
-
-# Linux
-*.swp
-
-# Windows
-Thumbs.db
-Desktop.ini
-
-# IDEs and Editors
-.vscode/
-.idea/
-*.sublime-project
-*.sublime-workspace
-*.tmproj
-
-# Logs and Debugging
-*.log
-*.out
-*.pid
-
-# Temporary files
-*.tmp
-*.bak
-
-# Secrets and Environment Variables
-.env
-.env.local
-.env.development
-.env.test
-.env.production
-
-# Miscellaneous
-coverage/
-*.lcov
-*.lock
