--- backend/requirements.txt
@@ -1,8 +1,0 @@
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-pydantic==2.5.0
-google-generativeai==0.3.2
-sqlalchemy==2.0.23
-psycopg2-binary==2.9.9
-python-dotenv==1.0.0
-python-multipart==0.0.6
