--- cloudbuild-frontend.yaml
@@ -1,44 +1,0 @@
-steps:
-  # Step 1: Authenticate Docker to Artifact Registry for us-west1
-  - name: 'gcr.io/cloud-builders/gcloud'
-    args: ['auth', 'configure-docker', 'us-west1-docker.pkg.dev']
-
-  # Step 2: Build the frontend Docker image
-  - name: 'gcr.io/cloud-builders/docker'
-    args: [
-      'build',
-      '-t', 'us-west1-docker.pkg.dev/$PROJECT_ID/ozlistings-frontend/frontend:$BUILD_ID',
-      '-f', 'frontend/Dockerfile',
-      './frontend'
-    ]
-
-  # Step 3: Push the frontend Docker image to Artifact Registry
-  - name: 'gcr.io/cloud-builders/docker'
-    args: [
-      'push', 'us-west1-docker.pkg.dev/$PROJECT_ID/ozlistings-frontend/frontend:$BUILD_ID'
-    ]
-
-  # Step 4: Deploy frontend to Cloud Run in us-west1
-  - name: 'gcr.io/cloud-builders/gcloud'
-    args:
-    - 'run'
-    - 'deploy'
-    - 'ozlistings-chat-frontend'
-    - '--image'
-    - 'us-west1-docker.pkg.dev/$PROJECT_ID/ozlistings-frontend/frontend:$BUILD_ID'
-    - '--region'
-    - 'us-west1'
-    - '--platform'
-    - 'managed'
-    - '--allow-unauthenticated'
-    - '--port'
-    - '80'
-    - '--set-env-vars'
-    - 'REACT_APP_BACKEND_URL=https://ozlistings-chat-876384283449.us-west1.run.app'
-
-images:
-  - 'us-west1-docker.pkg.dev/$PROJECT_ID/ozlistings-frontend/frontend:$BUILD_ID'
-
-options:
-  logging: CLOUD_LOGGING_ONLY
-  defaultLogsBucketBehavior: REGIONAL_USER_OWNED_BUCKET
