--- docker-compose.backend.yml
@@ -1,29 +1,0 @@
-version: '3.8'
-
-services:
-  backend:
-    build:
-      context: ./backend
-      dockerfile: Dockerfile
-    ports:
-      - "8001:8000"
-    environment:
-      - PORT=8000
-      - DB_USER=${DB_USER:-postgres}
-      - DB_PASSWORD=${DB_PASSWORD:-password}
-      - DB_HOST=${DB_HOST:-localhost}
-      - DB_NAME=${DB_NAME:-ozlistings}
-      - FRONTEND_URL=${FRONTEND_URL:-http://localhost:3000}
-      - GEMINI_API_KEY=${GEMINI_API_KEY}
-    volumes:
-      - ./backend:/app
-    restart: unless-stopped
-    healthcheck:
-      test: ["CMD", "python", "-c", "import requests; requests.get('http://localhost:8000/')"]
-      interval: 30s
-      timeout: 10s
-      retries: 3
-      start_period: 40s
-
-# Note: This configuration is for backend-only local development
-# Production backend is deployed at: https://ozlistings-chat-876384283449.us-west1.run.app
