--- frontend/nginx.conf
@@ -1,12 +1,0 @@
-# frontend/nginx.conf
-server {
-    listen 8080;
-    server_name localhost; # Or your domain if you have one
-
-    root /usr/share/nginx/html;
-    index index.html;
-
-    location / {
-        try_files $uri $uri/ /index.html;
-    }
-}
