--- frontend/package.json
@@ -1,16 +1,0 @@
-{
-    "name": "ozlistings-frontend",
-    "version": "1.0.0",
-    "private": true,
-    "dependencies": {
-      "react": "^18.2.0",
-      "react-dom": "^18.2.0",
-      "react-scripts": "5.0.1",
-      "react-markdown": "^8.0.7",
-      "serve": "^14.2.1"
-    },
-    "scripts": {
-      "start": "react-scripts start",
-      "build": "react-scripts build"
-    }
-  }
