--- frontend/src/App.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import EnhancedChat from './components/EnhancedChat';
-
-function App() {
-  return <EnhancedChat />;
-}
-
-export default App;
