--- pyproject.toml
@@ -1,22 +1,0 @@
-[tool.poetry]
-name = "ozlistings-ai-agent"
-version = "0.1.0"
-description = "AI Agent for Ozlistings chat"
-authors = ["Your Name <you@example.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.9"
-fastapi = "^0.100.0"
-uvicorn = "^0.20.0"
-psycopg2-binary = "^2.9.5"
-sqlalchemy = "==1.4.47"
-pydantic = ">=2.8.0,<3.0.0"
-google-generativeai = "*"
-python-dotenv = "^1.0.0"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^7.1.2"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
